Drop infinitely recursive assignment operators
CelBoolVar and CelNumVar had incorrect implementations
for their assignment operators. As they lead to infinite
recursion we can be sure they were never actually called
anywhere. However they still are instantiated in the tests
so we have to change the CelExpression definition to accept a
const argument to make things compile again.